creating a DistributionService and minor changes
