--- TelegramBot/TelegramBot/Messages/BusyUsers.cs
@@ -1,9 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace TelegramBot.Messages
-{
-    public static class BusyUsers
-    {
-        public static List<long> BusyUserId { get; set; }
-    }
-}
